refactor: remove LSP depends in ts pkgs
