+++ extensions/vscode-vue-language-features/package.json
@@ -634,13 +634,12 @@
 		"@volar/shared": "0.33.2",
 		"@volar/vue-language-server": "0.33.2",
 		"@vue/compiler-dom": "^3.2.27",
+		"@vue/compiler-sfc": "^3.2.27",
 		"@vue/reactivity": "^3.2.27",
 		"esbuild": "latest",
 		"upath": "^2.0.1",
 		"vsce": "latest",
-		"vscode-html-languageservice": "^4.2.1",
 		"vscode-languageclient": "^8.0.0-next.8",
-		"vscode-languageserver-textdocument": "^1.0.3",
 		"vscode-nls": "5.0.0"
 	}
 }
