+++ extensions/vscode-vue-language-features/src/features/splitEditors.ts
@@ -1,10 +1,7 @@
 import * as vscode from 'vscode';
 import { ref, computed } from '@vue/reactivity';
 import * as shared from '@volar/shared';
-import * as html from 'vscode-html-languageservice';
-import { TextDocument } from 'vscode-languageserver-textdocument';
-
-export const htmlLs = html.getLanguageService();
+import { parse, SFCBlock } from '@vue/compiler-sfc';
 
 export function activate(context: vscode.ExtensionContext) {
 
@@ -18,7 +15,7 @@ export function activate(context: vscode.ExtensionContext) {
 		if (!editor) return;
 
 		const doc = editor.document;
-		const descriptor = getDocDescriptor(doc.getText());
+		const { descriptor } = getDocDescriptor(doc.getText());
 		const leftBlocks = [
 			descriptor.scriptSetup,
 			descriptor.script,
@@ -33,18 +30,18 @@ export function activate(context: vscode.ExtensionContext) {
 		await vscode.commands.executeCommand('workbench.action.toggleSplitEditorInGroup');
 		await foldingBlocks(rightBlocks);
 
-		async function foldingBlocks(blocks: shared.SfcBlock[]) {
+		async function foldingBlocks(blocks: SFCBlock[]) {
 
-			const firstBlock = blocks.sort((a, b) => a.startTagEnd - b.startTagEnd)[0];
+			const firstBlock = blocks.sort((a, b) => a.loc.start.offset - b.loc.start.offset)[0];
 
 			const editor = vscode.window.activeTextEditor;
 			if (!editor) return;
 
-			editor.selections = blocks.map(block => new vscode.Selection(doc.positionAt(block.startTagEnd), doc.positionAt(block.startTagEnd)));
+			editor.selections = blocks.map(block => new vscode.Selection(doc.positionAt(block.loc.start.offset), doc.positionAt(block.loc.start.offset)));
 
 			await vscode.commands.executeCommand('editor.unfoldAll');
 			await vscode.commands.executeCommand('editor.foldLevel1');
-			editor.revealRange(new vscode.Range(doc.positionAt(firstBlock.startTagEnd), new vscode.Position(editor.document.lineCount, 0)), vscode.TextEditorRevealType.AtTop);
+			editor.revealRange(new vscode.Range(doc.positionAt(firstBlock.loc.start.offset), new vscode.Position(editor.document.lineCount, 0)), vscode.TextEditorRevealType.AtTop);
 		}
 	}
 }
@@ -52,7 +49,7 @@ export function activate(context: vscode.ExtensionContext) {
 function useDocDescriptor() {
 
 	const splitDocText = ref('');
-	const splitDocDescriptor = computed(() => shared.parseSfc(splitDocText.value, htmlLs.parseHTMLDocument(TextDocument.create('', '', 0, splitDocText.value))));
+	const splitDocDescriptor = computed(() => parse(splitDocText.value, { sourceMap: false, ignoreEmpty: false }));
 
 	return getDescriptor;
