+++ packages/vue-code-gen/package.json
@@ -17,7 +17,6 @@
 	},
 	"dependencies": {
 		"@volar/code-gen": "0.33.2",
-		"@volar/shared": "0.33.2",
 		"@volar/source-map": "0.33.2",
 		"@vue/compiler-core": "^3.2.27",
 		"@vue/compiler-dom": "^3.2.27",
