+++ packages/vue-tsc/package.json
@@ -16,7 +16,6 @@
     "vue-tsc": "./bin/vue-tsc.js"
   },
   "dependencies": {
-    "@volar/shared": "0.33.2",
     "@volar/vue-typescript": "0.33.2"
   },
   "peerDependencies": {
