+++ packages/vue-typescript/src/use/useSfcStyles.ts
@@ -1,12 +1,11 @@
 import { computed, Ref } from '@vue/reactivity';
 import * as SourceMaps from '@volar/source-map';
-import * as shared from '@volar/shared';
-import { Embedded, EmbeddedFile } from '../vueFile';
+import { Embedded, EmbeddedFile, Sfc } from '../vueFile';
 import { EmbeddedFileSourceMap } from '../utils/sourceMaps';
 
 export function useSfcStyles(
 	fileName: string,
-	styles: Ref<shared.Sfc['styles']>,
+	styles: Ref<Sfc['styles']>,
 ) {
 
 	const files = computed(() => {
