+++ packages/vue-typescript/src/use/useSfcTemplate.ts
@@ -1,12 +1,11 @@
-import * as shared from '@volar/shared';
 import { computed, Ref } from '@vue/reactivity';
 import { EmbeddedFileSourceMap } from '../utils/sourceMaps';
 import * as SourceMaps from '@volar/source-map';
-import { Embedded, EmbeddedFile } from '../vueFile';
+import { Embedded, EmbeddedFile, Sfc } from '../vueFile';
 
 export function useSfcTemplate(
 	fileName: string,
-	template: Ref<shared.Sfc['template']>,
+	template: Ref<Sfc['template']>,
 ) {
 
 	const file = computed(() => {
