+++ packages/vue-typescript/src/utils/moduleSpecifierCache.ts
