+++ packages/vue-typescript/src/utils/packageJsonCache.ts
